Initial commit of an OpenGL Transmission Format (glTF) BIMServer serializer plugin in conjunction with the following ticket: #105.
This commit provides medium changes to the existing procedure of the
Collada serializer on which the OpenGL Transmission Format depends. This
dependency is in the form of the collada2gltf exporter presented by
Khronos.

"collada2gltf" is required to be on the PATH/available to the
command-line at the operating system level so that the plugins can
function.

The OpenGL Transmission Format family of plugins can and should be
consolidated into 1 single plugin if/when plugin configurations can be
made from the plugin.xml file.

Conflicts:
	Collada/src/org/bimserver/collada/ColladaSerializer.java