Issue #1353, fixed NullCall in GetProjectByUuidDatabaseAction.java
In this commit, a new query method was added to DatabaseSession.java
which returns a hashmap of an IdEObject value with a UUID key (this
allows for a faster run time via using HashMap.contains()
and allows for the use of this query with other IdEObjects).
In GetProjectByUuidDatabaseAction.java, logic was added to
execute the new queryUuid() method while also performing
authorization checks.

Reported-By: zaqifathis